musicPro/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.1
django-bootstrap4==23.1
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
